Add links to article/single call
@korsvanloon you may now call /articles/single?url= with any url and you'll allways
get a { links: [] } object which contains a request-link with the correct url to
post a request to. You may always provide the article_url : windows.location.href.
This way it does not matter which url you call.

Whether or not the response contains a { requests: [] } and/or { annotations: [] } should
trigger rendering of requests / annotations.

You will need to adapt your AJAX code to also deliver content when the response code is
different from 200, since you'll get a 404 with valuable content.